requirements.txt
openai
numpy
pandas
scikit-learn
flask
python-dotenv